Introduce `ContentVariant` to the database-adapter
The content variant defines _how_ content is maintained, whether it is stored
only in a Nessie commit (on reference) or whether it requires global state.

`ContentVariant` currently only distinguishes the two kinds above. It allows
the `DatabaseAdapter` implementations to do the things necessary for a
specific variant - e.g. only handle, especially read, global state for
variants that require global state.

There will be a 3rd variant for the variant that stores Iceberg metadata in
Nessie.