Create Exception Mapper that specifically handles ConstraintViolationExceptions
The reason to do this is that it's possible that users register their
own exception mappers that handle `ConstraintViolationExceptions`. Given
that our current `ValidationExceptionMapper` is registered on the
superclass of `ConstraintViolationExceptions`, namely
`ValidationException`, it's possible that the exception handling code
for `ConstraintViolationException` in `ValidationExceptionMapper` isn't
called. This is because during exception validation, a distance for the
current exception is calculated to the exception that is being handled
by a particular exception mapper.

For example, here's how Glassfish determines the exception mapper to use
when an exception happens: https://github.com/eclipse-ee4j/jersey/blob/master/core-common/src/main/java/org/glassfish/jersey/internal/ExceptionMapperFactory.java#L111-L131

Notice the distance calculation in
https://github.com/eclipse-ee4j/jersey/blob/master/core-common/src/main/java/org/glassfish/jersey/internal/ExceptionMapperFactory.java#L214-L226.
This means that our current `ValidationExceptionMapper` with a distance
of 1, whereas the `ConstraintViolationExceptionMapper` would end up with
distance=0 for a `ConstraintViolationException`