Nessie SQL extensions: Fix Use reference at hash (#3440)
for use reference @ hash sql command, we are calculating/validating the ref+ hash. But in the end we are not setting hash to the catalog. only ref name is set to the catalog. Hence the issue.

These are changes from Nessie side for #3439
Once this is merged and release is made, need to fix at Iceberg side using this configuration.
Later end to end test has to be updated in ITNessieStatements once Iceberg releases it.