Shave off two datbase reads reads per Nessie operation
To write a new global log entry, Nessie reads the parent hashes from the current global log entry.
This can be avoided by storing the parent hashes in the `GlobalStatePointer` as well (nothing to do for transactional databases).

To write a new ref log entry, Nessie reads the parent hashes from the current ref log entry.
This can be avoided by storing the parent hashes in the `GlobalStatePointer` as well (transactional database: add the parent hashes to `TABLE_REF_LOG_HEAD`).

The change can easily be implemented to be backwards compatible:
* If the list or parent hashes is not present, read the parent hashes like the current implementation does.
* If the "head" of the parent hashes is not equal to the expected "head", a concurrently running Nessie server updated the `GlobalStatePointer` (or `TABLE_REF_LOG_HEAD`) -> fall back to the old behavior and read the parent hashes like the current implementation does.

Fixes #3448