Use Immutable Builder functionality in HTTP Param classes to generate Builders
The only reason we can't fully move those classes to true Immutables is
because of bean validation requiring those classes to have a default
constructor. Note that bean validation on those classes is done via
`@BeanParam`, such as in `getXyz(@BeanParam EntriesParams params)`.

At the very least, we can remove the hand-written builders and let the
Immutables lib do the work of generating those (since we already have
tests in place that make sure objects are properly constructed and
validated).