Remove unused 'type' + "type-extractor" functions
The `type` field stored in the global-log was never really used and also always `0`.

This change removes the `type` field from the immutable-types, renames it in the serialized representation, and removes the "type extractor" arguments from the respective functions in `DatabaseAdapter`.

Note: the `type` field for "on reference" contents / content keys is correct and used.