`AbstractDatabaseAdapter.fetchValues()` looking for non-existing keys
A `ContentApi.getContents()` with keys that do not exist leads into
traversing the whole commit log.

The current code in `AbstractDatabaseAdapter.fetchValues()` is optimized to stop
traversing the commit log while looking for content-values when the values for
all requested keys have been found.

In turn this means that the commit-log-traversal will look into _every_ commit
for keys that do not exist.

It's is rather easy to prevent that unnecessary whole-commit-log-traversal for
non-existing keys: Every 20th (by default) commit contains the whole key-lists,
so all visible (think: existing + undeleted) keys. The set of all keys can be
inspected to stop looking for keys that are known to not exist.

While it sounds like a minor optimization, it's in fact a rather big one.