Record commit-id in key-lists (#3592)
* Record commit-id in key-lists

Fetching content values can become expensive as fetching the value from a
`Put` opartion may require reading a lot of commit-log-entries, the more
commit-log-entries the older the content.

This change adds the commit-id of the last `Put` operation to the already
existing entries in `KeyList`, which allows `DatabaseAdapter.fetchValues`
to quickly "navigate" to the commits that contain the most recent `Put`
operation that contains the content value for the requested content key.

Time/resource complexity for existing keys is reduced from "number of
commit log entries from HEAD to last Put operation" to at max two
bulk-fetch operations against the commit-log.

Updates & renames the type `KeyWithType` to `KeyListEntry`, adds the
commit-id to `KeyListEntry`.

This is not a breaking change and requires no data migration. Upgrade
tests are included. Although not tested, and therefore not guaranteed,
this change should also be safe to use during a rolling-upgrade.

Fixes #3591

* update javadoc

* Dynamically calculate commits-per-version