Make sure Namespace/ContentKey don't have empty strings in between
This also required some changes to `NessieLogStore` because that part of
the code would simply do a `path.split("/")` where the first element
would then end up being an empty string if `path` started with a `/`.

The error messages have also been improved because if the check fails,
then it's difficult to tell exactly why it failed.