Add content-id to `BatchAccessChecker.canReadContentKey()`
Nessie functionality to list all keys can now use the content-id in addition to the content-key for authorization.
Nessie functionality to filter the optionally returned put operations when listing the commit-log can now use the content-id in addition to the content-key for authorization, delete operations will not have the content-id.

Moves the authorization part for `TreeApi.getEntries` to `TreeApiImplWithAuthorization.filterEntries()`.

Also removes the duplicate call to `namedRefWithHashOrThrow()` from `getEntries()` + `getCommitLog()`.