Store GC identify action's results into a table (#3421)
Currently, output of `GCImpl#identifyExpiredContents` is stored in an in memory map by merging the results from each spark task.

The results can be stored in a table instead, to run clean up afterwards or to be used by consumer of a different process. 

With this PR, Spark is directly writing the task output into the iceberg table in a configured reference where each task's output will be mapped to multiple rows in a table (one row per content id).

`GCImpl#identifyExpiredContents` now returns the run id of the completed task. Using which user can query the output table. Also added an interface to fetch the latest completed run id.

All old gc test cases now query the output table to collect the results and validates it.
Added a new test case to test in-progress and marker row write to the table.

Also removed `Serializable` from `Content` along with removing intermediate classes (`contentValues`,`IdentifiedResult`)that was used for storing output in-memory map.