Squash-merge
Updates the merge-operation to squash the commits to be merged, if there is more than 1
commit to be merged.

The most recent change to any content-key will be present in the squashed commit,
intermediate changes will not be present in the squashed commit.

Commit messages will be concatenated with a separating line consisting of a sequence of
`-` characters.

Commit properties are combined, last property value will be present in the squashed
commit.

Commit author will be reset to the user issuing the merge operation in a squashed
commit. Commit author-time will be reset to the current timestamp for a squashed commit.

This API-change is not changing the signature, but the behavior changes, as the default
is then to squash the commits to merge. I.e. not specifying the `keepIndividualCommits`
parameter for a merge operation means "squash" the commits. To use the previous behavior,
pass `keepIndividualCommits=true`.

This change allows `MERGE` via SQL like this:
```sql
MERGE dev_branch INTO main
  ONLY TABLE only.table.to.merge;
```
or
```sql
MERGE dev_branch INTO main
  EXCEPT TABLE not.this.table;
```
or
```sql
MERGE dev_branch INTO main
  FORCE TABLE blindly.apply.this.table.onto.main;
```

Fixes #3809