Remove global state for Iceberg tables + views
Before this change, the location to the Iceberg table/view metadata was maintained
globally, so all commits/references/branches/tags shared the same metadata.
It was originally introduced to ensure that there can be no duplicate IDs for
column-ID, schema-ID, et al. On the flip side, the global metadata "leaked" all
table snapshots and schemas to all branches and even experimental branches could
(potentially) exhaust all possible column IDs for a table.

So this change moves the Iceberg metadata-pointer back to "per commit".