Use group separator (u001D) character instead of zero byte for Namespace/ContentKey (#3850)
For backward-compatibility we still allow the zero byte, but convert it
to the group sep. character, since that's our new canonical representation
of a "." delimiter within a single element name.

Note that this also means that
`Namespace.parse("a.b\u001Dc.d")` is the same as `Namespace.parse("a.b\u0000c.d")`.

The reason for changing from the zero byte to the group sep. character is
because certain webservers (Jetty) don't allow the zero byte to be used
in the URL path
(https://github.com/eclipse/jetty.project/blob/jetty-10.0.x/jetty-http/src/main/java/org/eclipse/jetty/http/HttpURI.java#L1203-L1206).
This means that the mentioned path check in Jetty fails even before for
example the `NamespaceParamConverter` gets called. Due to that we
decided to switch our canonical representation of a ContentKey/Namespace
to use a new character instead of the zero byte.