Do not write empty global-log-entries
The current non-transactional database-adapters write empty global-log entries
for changes that do not affect global state.

Since it is not great to write empty global-log-entries, this change
optimizes the non-transactional database-adapter implementation to no longer
write "empty" global-log entries. The change there is to split the global-id,
which was used to both point to the head of the global-log and to serve as the
reference value for the global-pointer CAS operation. The head of the global
log is now maintained as a separate field in global-pointer so that the global-id
can still be used as the reference value for the CAS operation.

Prerequisite for #3866