Let StoreWorker pull global-state
`StoreWorker` implementations previously got an `Optional` containing the
value of a content's global state. This change changes the behavior to
pull the global state value.

Also marks an attribute for delta-lake-tables as optional to properly handle
the case coded in `TableCommitMetaStoreWorker`.

`PersistVersionStore` checks the content-IDs of the new and expected values
in a `Put` operation.

Migrates to AssertJ for tests in `servers/store`.

Prerequisite for #3866