Migrate to `@QuarkusIntegrationTest` (#3988)
Quarkus improved and simplified integration testing, especially against native images,
`@NativeImageTest` is deprecated now.

This change migrates all integ-tests to `@QuarkusIntegrationTest` and removes the now
superfluous `ITNative*` test classes and test profiles.

Running the tests "with Docker enabled" (using the `-Pdocker` profile, which configures
`quarkus.container-image.build=true`) lets the Quarkus application (i.e. Nessie server)
to test run in a Docker container with a custom, dynamic Docker network. So all
test containers (Postgres + Dynamo) need to "join" the same Docker network. But it also
means that starting a process on the host (think: mongod via flapdoodle) does not work
with that approach, hence updated the `MongoTestResourceLifecycleManager` to use
testcontainers as well.

As a not-so-great consequence, `ITOpenIdAuthentication` cannot run with `-Pdocker`,
because the `OidcWiremockTestResource` runs in the JVM running the tests on the Docker
host and is not accessible from the Docker container running the tested Nessie server.

See https://quarkus.io/guides/getting-started-testing#native-executable-testing