Generate content ID on the server side & namespace content-IDs
This change is driven by the demand to have "unique" content-IDs for
namespaces. Technically there are two kinds of namespaces:
* explicitly created namespaces (via `NamespaceApi.createNamespace`)
  and
* implicitly "created" namespaces, which are returned by
  `NamespaceApi.getNamespaces`, if no explicitly created namespace
  exists.

The previous approach had a deterministic content ID, because
content ID == namespace-path. "Unique" (aka `UUID.randomUUID()`)
are not deterministic.

To be able to distinguish between implicitly and explicitly created
namespaces (think: retrieving an implicitly created one, adding
properties and persisting it vs. the same with an explicitly created
namespace), the content ID must be `null` for implicitly created
namespaces and non-`null` for explicitly created ones.

This mandates the change to make `Content.getId()` `@Nullable` and
remove the "implicit" creation of a content ID, if not present.

Also implicitly changes the technical behavior that content-IDs are
now _only_ created on the server side. Previously, the content-ID
was generated for clients using the Java `:nessie-model` classes.

A content object to be created must have no (== `null`) content ID
and the expected-content object must be `null` as well.
A content object to be updated must have a non-`null` content ID,
and the expected-content object must be non-`null` as well, with
the same content-ID.

Fixes #3731
Fixes #4017