Add content-variant using attachments
Allow the `StoreWorker` to handle a third "content variant" in addition to "content on reference" and "content with global state". The third variant is actually an extension to "content on reference" and allows attachments. Each commit stores references to the required attachments, which are, in case of Iceberg, the "shallow metadata" and all child objects. Note: This technical change is not specific to Iceberg and a purely internal concern.

(See point 2 in #4138)