Use hashOnRef when fetching Namespaces
Currently the `hashOnRef` query flag was not used when fetching a
Namespace at a particular hash. This diff makes sure that the
`hashOnRef` flag is properly used when fetching namespaces.
Note that we don't use `hashOnRef` when creating/updating/deleting a
Namespace and thus pass a `null` for the hash param to `namedRefWithHashOrThrow(...)`