Escape all column names in SQL DML+DDL statements
`org.projectnessie.versioned.persist.tx.SqlStatements` uses unescaped column names.
This can and will lead to issues with certain database products and versions,
as it happend with the implicit version bump of H2 from 1.4.x to 2.1., which "sneaks
in" via the Quarkus version bump to 2.9.0, where `value` became a keyword.