Remove the configurable default for the configurable values
The max entity size for the (embedded) key list is configured using two values from the
`DatabaseAdapterConfig` interface via `getMaxKeyListSize` and `getDefaultMaxKeyListSize`,
the latter serves as the default for the former. This is confusing and actually wrong.

This change removed the configurable default entirely and introduces an optional
"hard limit" enforced by database adapter implementations - DynamoDB configured to 375kB.