Identify referenced+unreferenced HEADs
Implement an efficient way to retrieve all unreferenced HEAD commits.

As a side effect (and up for discussion whether the information's useful), it also returns
* all fork-points (commits that are the parent of more than one other commits)
* map of commit-ID to named-refs

The implementation scans all commits and identifies all "heads" (aka commits that are not the parent of another commit) and all fork-points (commits that are the parent of more than one other commits).
In a next step all named-references and their commit-logs are scanned to identify the HEADs that are unreferenced and those that are the HEADs of named references.

"Efficient" means the only somewhat expensive part is the scan of all commits, but the computation/identification of all HEADs and fork-points (commit with more than one "child") plus the differentiation of unreferenced & referenced HEADs is not that expensive. Plus: it's a server-side implementation and interested clients (maintenance + GC) could just ask for this information.