Add Rolling-Upgrade test suite
Adds a new integration-tests for rolling upgrades including the
"machinery" to run rolling-upgrade tests.

That "machinery", implemented in `RollingUpgradesExtension` uses
MongoDB as the backend database. Using RocksDB doesn't work, because
rolling-upgrades requires that two server instances access the same
database - and RocksDB doesn't allow that.

Tests using `RollingUpgradesExtension` get two instances of
`NessieApiV1` - one configured against the "old" Nessie version and
the other against the "current" (in-tree) Nessie version.

`RollingUpgradesExtension` reinitializes the Nessie repository for
each combinations of Nessie versions.

The new `ITRollingUpgrades` tests a couple scenarios, against both
the old and new versions:
* get the default branch
* create new branch
* verify created branches
* commit a new table
* verify contents

`ITRollingUpgrades` is configured to require at least Nessie 0.31.0,
which is not yet released, because any rolling upgrade to current
in-tree will fail, because it is not safe for a rolling-upgrade
since #4234.

Fixes #4348