Store and retrieve merge-parent (#4333)
Merge operations store the merged commit ID as an additional parent.

This change introduces a new field `additionalParents`, which is a
list of commit-IDs. This list is usually empty, except for
squashing merge commits, that are not single-commit fast-forward
merges.

Transplant operations do _not_ store an "additional parent".

Merge operations that are configured to keep separate commits do
_not_ store an "additional parent" as well, because this is effectively
the same as transplanting (cherry-picking).

This change _prepares_ the OpenAPI type `LogEntry` with a new attribute
`additionalParents`. The attribute will _not_ be populated though to
not break older clients. Otherwise, older clients will fail to
deserialize commit log entries.

This change however does return the first additional parent as a
property in `CommitMeta`, with the property key `_merge_parent`, so
clients that know about the meaning of that property, can handle it.