versioned/spi: remove `Collator` in `Key`
`java.text.Collator` is nice when working with real text, but content
keys are not "real text" (with for example german umlauts, french
accents, etc).

`Collator` is used in `org.projectnessie.versioned.Key` for all kinds
of preformance critical operations like `equals()`, `hashCode()` and
`compareTo()`.

Expecially #4536 suffered due to the use of `Collator`, which became
obvious in the test case `TestKeyListBuildState#manyBuckets`, because
it was super slow and a flame graph immediately pointed to the use
of `Collator`. A single iteration of that test takes ~2ms locally
in IntelliJ without `Collator`, but 70ms with `Collator` for 500 `Key`s.

(The change in AbstractDatabaseAdapter is rather a nice-to-have and
a unit test fix.)