Hashed key-lists using open-addrressing (#4536)
Refactors key-lists handling in `CommitLogEntry`.

The `KeyListEntry`s in the key-list(s) in `CommitLogEntry` are currently organized by "least recently updated keys first", which degrades read performance with many keys per reference.

This PR re-organizes the keys using an open-addressing hash-map, with the content's `Key` as the open-addressing hash-map's key. The whole open-addressing hash-map is split into multiple segments: the first segment is the "embedded" key-list inside `CommitLogEntry`, with additional segments as `KeyListEntity`s. This allows effective point-queries, both to only fetch the required `KeyListEntry` and within the `KeyListEntry`.