Add build used by integrations-testing for Iceberg dependencies
Builds on the changes for #4645 and adds a usually unused Gradle
build in the `iceberg/` directory. This new Gradle build will be
used by integrations-testing to build the Nessie projects that
have dependencies to Apache Iceberg.

Tools integrations testing will require a "surrounding" Gradle
build, which in turn includes three builds:
* Nessie build, but without all projects that depend on Apache
  Iceberg artifacts. For example: nessie-gc-base, spark-*
* Apache Iceberg build, which then uses the artifacts of the Nessie
  build.
* Nessie-Iceberg build containing the Gradle projects that are
  excluded in the first Nessie build, which can that way reference
  both the Nessie and Apache Iceberg artifacts. This includes the
  nessie-gc-base and nessie-spark-* projects.