Move antlr-runtime dependency back into Nessie
Back in the (Maven) days, we had to move the shaded antlr-runtime to
a separate build, because otherwise the non-shaded and shaded
antlr4-runtime classes conflicted and the build became quite tricky.

Now with Gradle it's possible to get rid of the "artificial"
nessie-antlr-runtime dependency using the same group-id but a
different version schema.

Adding back the relocated antlr4 runtime as a nessie project.

This eliminates the need for https://github.com/projectnessie/nessie-antlr-runtime/