Handle times in SQL extensions better fixes #1452
We start accepting dateTime with time zone only. If the user passes the dateTime without zone:
"2022-06-29T09:26:07.974992" we should throw an exception denoting that the user needs to pass the zone explicitly. The new dataTime format should be: 2022-06-29T11:26:07.976777+02:00. It means that we are breaking the compatibility of the current API, but we concluded that having an exception with a meaningful message would be enough in that case. Also, Nessie is not 1.x version yet, so the API changes can be justified.