Let SQL `CREATE` and `ASSIGN` use `USE`d ref
The current behavior of `CREATE BRANCH|TAG` is to fall back to the
default branch, if the `FROM` clause is omitted.
The current behavior of `ASSIGN BRANCH|TAG` is to fall back to the
default branch, if the `TO` clause is omitted.

As outlined in #4685, the above behavior is not intuitive.

The PR changes the above behaviors to fall back to the name of the
reference configured for the Spark session, honoring a `USE`
statement issued by the user.

Fixes #4685