Nessie REST (HTTP) client: verify JSON response
When users configure the REST base URI incorrectly, e.g. by setting
it to `http://127.0.0.1:19120/` against a locally running
Nessie-Quarkus-Server, they just get some "cannot parse response
entity" error message, which isn't helpful.

This change verifies the the `Content-Type` returned by the server
is a valid JSON one, one that ends in either `/json` or `+json`.
If the response is not a JSON response, the thrown exceptions gives
a better hint to what _could_ be wrong.