Avoid converting 415 Unsupported Media Type into 500 ISE (#4436/#4711)
In cases where a client request specified a Content-Type that a
Nessie-Quarkus method wouldn't accept, NessieExceptionMapper would
convert the resulting 415 exception into a generic 500 exception.  The
reason string still provided a useful hint ("Cannot consume content
type"), but the status code was more general than necessary.

This commit adds a case to JAX-WS NotSupportedException case to
NessieExceptionMapper, preventing it from falling through to the
catch-all 500 ISE case.  Incorporating review feedback, this commit
does so with a new subclass of NessieRuntimeException. It also
adds a couple of tests covering the change.