Nessie GC: mark & sweep (`gc-base` module)
See also the readme in `./gc/README.md`.

GC is implemented as a two-phase approach:
1. Identify all refetences to live content versions by walking all
   named references. Content versions for Iceberg are tuples of
   commit-ID + snapshot ID + content ID + metadata reference).
   These `ContentReference`s are and stored in an instance of
   `LiveContentSetsStore`. This is the "mark phase". The set of
   content-references is called "live contents set".
2. For each content-ID, resolve all `ContentReferences` to file
   objects and populate a bloom-filter with those. Then traverse
   the base storage locations for each content-ID and delete all
   files that are not contained in the bloom filter, but do not
   delete files that are "too new" (created since the "mark phase"
   started).

A command line tool `nessie-gc-shell` is provided as well. It
supports running mark-and-sweep, mark only, sweep only plus a few
supplemental commands.

All pieces are pluggable, there are Java interfaces (or abstract
classes) providing the API for all pieces. This makes the code much
easier to test and helps using alternative implementations.

The base functionality is contained in the `nessie-gc-base` project.
This has default implementations to identify live content
references, an in-memory live-content-sets-storage and a default
implementation to run expire per content.

Support for content types, like Iceberg tables, implements
the functionality to expand content-references to actual
file-objects, list file-objects from base locations and delete
files. The "sweep" phase for Iceberg therefore also acts as the
"Nessie aware delete orphan files".

A live-content-sets-storage implementation using JDBC is present as
well.  Small Nessie repositories using a "one-off" mark-and-sweep
run can get away with the in-memory contents-storage. But big Nessie
repositories would require too much heap.

Another reason to eventually use a persistent
live-content-sets-storage is to allow the deletion of the files
occupied by completely unreferenced tables - tables that are not
referenced by any live Nessie commit.  This functionality is not
implemented, but would work by collecting the base-locations for
all content-IDs that are no longer present in the most recent
live-contents-set.

This PR still has a few non-`Test*` classes that were only added to
prove that the concept works quick and does not require much heap
and does not use a lot of CPU. Those "tests" in `nessie-gc-base` and
`nessie-gc-iceberg-inttest` will be removed, because those add no
real value and were only needed to validate the concept. The "final"
proof showed that a mark-and-sweep (deleting ~1.7M files, collected
from "many" branches, commits, content objects) alone can complete
in round about one minute with a 2GB Java heap.

Fixes #4695 #4694 #4626 #2255 #706