IDE: Also detect Eclipse IDE (#4878)
IDE "detection" currently works fine for IntelliJ, but not for
Eclipse and related (VSCode) IDEs.

Gradle "command line" builds run for all Spark + Scala combinations,
reusing the same source folders for all Scala versions. But IDEs are
quite allergic when one source folder is used for multiple projects.
Therefore only one Scala version (usually 2.12) is "exposed" to IDEs.
This worked fine for IntelliJ, but did not for Eclipse.

This change adds Eclipse to the IDE "detection". (Sadly it has to be
duplicated in a few places.)