Expose name and full key to "get entries" filters
This will expose all of the available entry details to filters except
the content ID, which is an internal ID and probably should not be used
for filtering.

Rework the scrpt args in TreeApiImpl.filterEntries() to use CEL
wrappers.

Add proper CEL types for entry filter scrtips.

Drop VAR_NAMESPACE and VAR_CONTENT_TYPE decrarations that did not
have values even before this change.

Fixes #4912