Re-add `payload` to `StoreWorker` to identify the content-type
Previously we did not really enforce that the `payload` (a byte value
corresponding to a particular `Content.Type`) was correctly
persisted and as a workaround completely removed the handling of the
`payload` value. This is rather inefficient, because the
`StoreWorker` implementation has to deserialize the whole persisted
value to identify the content type.

Beside the above inefficiency having the right `payload` value is
also a prerequisite to make content-types completely pluggable.