Fix Quarkus logging during build (#4948)
Quarkus log messages are currently not properly formatted and look like this:
```
Detected a split package usage which is considered a bad practice and should be avoided. Following packages were detected in multiple archives: %s
```
but should really look like this:
```
WARN: Detected a split package usage which is considered a bad practice and should be avoided. Following packages were detected in multiple archives:
- "org.projectnessie.versioned.persist.mongodb" found in [org.projectnessie:nessie-versioned-persist-mongodb::jar, org.projectnessie:nessie-versioned-persist-mongodb:tests:jar]
```

Turns out that configuring the JBoss LogManager class was wrong, removing the configuration fixes the issue.