Change server reset method for python tests to not delete main branch (#5025)
Server reset method for python tests was based on deleting all branches. This would fail on newer servers as deleting main is not allowed anymore. The same result is achieved by resetting the main branch to its initial hash.

Changes are reflected in the reset_nessie_server_state method in conftest.py where all branches but main are now deleted and main is set to the no_ancestor hash.

Fixes #5024