Workaround OOMs during Quarkus test runs (#5075)
A recent PR CI run timed out, because of OOMs during
`:nessie-quarkus-cli:intTest`. This is reproducible locally, and
investigating some heap and thread dumps showed that the heap
pressure is actually caused by jacoco/class-instrumentation, which
reads many jar files. This might be a combination of Quarkus class
loading and jacoco instrumentation.

TL;DR the code added to `:nessie-quarkus-cli` caused more code to
instrumented, which led to the increased heap pressure, which in
turn caused the OOMs.

The default heap sizes for Quarkus tests is 256m..1024m. This
change bumps those values to 512m..1536m and moves that
configuration to `nessie-conventions`.

Note: the heap size for tests can also be controlled manually
using for example `-PtestHeapSize=2g` on the command line.