SQLExtension: Fix show reference command with use reference@hash (#5081)
Whenever SQL "USE REFERENCE @ HASH" is used, It only updates the NessieIcebergClient 's updatableReference. So, there is no change in the actual Nessie reference in the backend.

But when current reference info is fetched using "SHOW REFERENCE", It fetches from the backend which is not aware of "USE REFERENCE @ HASH". So, this is confusing for the users and also it is not the current state of the reference. Hence, this fix.

Assign ref, create ref from, and merge ref will honor the value set by "use reference @hash" now.