Simulate puts against large existing keysets (#5078)
This is related to, but does not implement, #3832.  Without adding
additional index structures, an exploratory implementation needed to add
a read-before-write on puts to support #3832.  The cost of the read was
sensitive to both the size of the set of keys present at the commit's
target branch, and to the number of puts in the commit.

This commit adds a simulation that exercises a related workload in a
simplified, single-user way.  The simulation could be generalized to
multiple users (it already uses a branch-to-Gatling-userid
correspondence that would support multiuser without changes).  However,
single-user was sufficient to explore the costs at the time.

The simulation might be more broadly useful regardless of what happens
with #3832.  This commit adds the sim without the exploratory impl it
was originally meant to test.  It also adds some
perftest/simulations/README.md hints for running individual simulations
with parameters, and adds a way to control the gatling loglevel from the
gradle commandline.