Isolate http-level test from java client-level tests (#5314)
Some tests in AbstractRestInvalidWithHttp used to construct HTTP requests
directly (by using an HttpClient). However, the majority of tests in the
"AbstractRest" suite rely on the Java client implementation to perform
actual HTTP requests.

Those tests are moved to AbstractResteasyTest for the sake of clarity and
uniformity of test approach and dependencies.

The now unused getHttpClient() is removed from AbstractRest. The unused
getUri() method is removed too.