Move Nessie Client construction into a JUnit5 extension
This is to unify Nessie Client handling in preparation for API v2 tests.

In general, this isolates the construction of the client, which is largely
dependent on the test env. from the test code itself, which operates at the
Nessie java client / API level.

NessieApiProvider is needed only to work around some @QuarkusTest oddities
as described in source comments in NessieClientResolver.

Deprecate but keep the old NessieApi annotation for compatibility with old
tests external to the Nessie repo.