Build UI against Nessie API v1 from 0.44.0
This is a temporary pin in preparation for API V2.

Newer builds will not publish OpenAPI YAML, which is why it is
necessary to refer to an older Nessie version in the UI build script.

The build uses a checked-in copy of Nessie 0.44.0 OpenAPI YAML file
at ui/src/openapi/

UI is to be updated to the upcoming V2 Nessie API eventually.