Extract service-side interfaces for client-facing services
This is in preparation for API v2.

Move `MergeBehavior` and `MergeKeyBehavior` to the top level.

The extracted interfaces (in the `services.spi` package) are intended
to be reused between API v1 and v2 REST endpoints.

Parameter holders are expanded into list of individual parameter
values to simplify reuse. Most of those parameter classes are going
to be different in API v2.

Return values types are preserved. V2 changes to `model` classes are
intended to be backward-compatible.