Move old API classes into the api.v1 package
This is in preparation for API v2, which is going to have parallel
set of HTTP interfaces and parameter classes using different
annotations.

Classes used by java clients are not affected. In this respect,
`FetchOption` could have been moved, but that would have broken
binary compatibility with older java clients. In fact `FetchOption`
probably belongs under the `model` package rather than the `params`
package, but it is already exposed to external clients, which is why
it remains in it old stop in this PR.

The moved classes are not expected to be referenced directly by
external code.