Allow `-` as a reference name in v2 REST paths
The literal minus char was meant to be used as a well-known
placeholder for the default branch in v2 REST paths.

However, #5004 missed those use cases. This PR fixes the path RegEx
to allow `-` as a reference specification and adjusts
RestV2TreeResource to treat it as the configured default branch name.

Fixes #5606