Support merge/transplant message overrides
Commit message may be overridden when the operation results in
exactly one commit on the target branch.

That is overriding messages in unsquashed transplants is not
supported.

Merge message overrides are supported only in API v2 at this point
because v1 does not have the commit message in merge HTTP payloads.

Closes #4430