Prevent using the Minio Extension on non-Linux OS
Background is that S3 URLs encode the bucket name in front of the S3
target's FQDN, which works fine for `bucket-name.localhost` on Linux,
but not on other OS'es. S3 requests encode the bucket name in the DNS
FQDN - e.g. `my-bucket.my-minio.com`, which translates to
`my-bucket.localhost` for local testing. Linux supports these
wildcards out of the box, but macOS + Windows do not.